Added a new parameter type to Matisse, CameraList.
This commit can be used as a reference on how to insert new parameter type and parameters.
Nota:

The only thing that is not visible in this commit is the fact that each time you create a new parameter in the dictionnary, you then have to run DictonnaryValidator, point it to MatisseWorkspace, click generate then publish and build again.